fix(vim): fix documentHighlight/codeLens only working on latest buffer
Since the autocmds are attached per buffer and we were overriding existing
autocmd groups by default, highlight and codelens only work on the latest
buffer. A way to fix this is to not specify an augroup. Instead, to
preserve the groups, we fixed this by not clearing the existing groups.
We just clear these commands when reattaching to the same buffer to
avoid duplicates.